frontend: Added option to skip repeated imperfect emulation warnings.
The option is called skip_warnings, and it must be set in ui.ini (it can
be set using the internal UI).  Red warnings cannot be skipped; yellow
warning can be skipped under certain circumstances.

For a yellow warning to be skipped, the system must have been launched
in a way that allows warnings to be displayed, in a configuration with
the same set of devices flagged with unemulated/imperfect features,
within the last seven days, and the warning must have been displayed
within the past 14 days.

Also fixed a bug with display of the MACHINE_NO_COCKTAIL flag in the
internal UI, and increased the size of XML integer attributes to 64
bits.

+++ src/lib/util/xmlfile.cpp
@@ -8,13 +8,12 @@
 
 ***************************************************************************/
 
-#include <cassert>
-
 #include "xmlfile.h"
 
 #include <expat.h>
 
 #include <algorithm>
+#include <cassert>
 #include <cctype>
 #include <cstdlib>
 #include <cstring>
@@ -582,7 +581,7 @@ const char *data_node::get_attribute_string(const char *attribute, const char *d
     found, return = the provided default
 -------------------------------------------------*/
 
-int data_node::get_attribute_int(const char *attribute, int defvalue) const
+long long data_node::get_attribute_int(const char *attribute, long long defvalue) const
 {
 	char const *const string = get_attribute_string(attribute, nullptr);
 	if (!string)
@@ -590,20 +589,20 @@ int data_node::get_attribute_int(const char *attribute, int defvalue) const
 
 	std::istringstream stream;
 	stream.imbue(f_portable_locale);
-	int result;
+	long long result;
 	if (string[0] == '$')
 	{
 		stream.str(&string[1]);
-		unsigned uvalue;
+		unsigned long long uvalue;
 		stream >> std::hex >> uvalue;
-		result = int(uvalue);
+		result = static_cast<long long>(uvalue);
 	}
 	else if ((string[0] == '0') && ((string[1] == 'x') || (string[1] == 'X')))
 	{
 		stream.str(&string[2]);
-		unsigned uvalue;
+		unsigned long long uvalue;
 		stream >> std::hex >> uvalue;
-		result = int(uvalue);
+		result = static_cast<long long>(uvalue);
 	}
 	else if (string[0] == '#')
 	{
@@ -690,7 +689,7 @@ void data_node::set_attribute(const char *name, const char *value)
     integer value on the node
 -------------------------------------------------*/
 
-void data_node::set_attribute_int(const char *name, int value)
+void data_node::set_attribute_int(const char *name, long long value)
 {
 	set_attribute(name, string_format(f_portable_locale, "%d", value).c_str());
 }
